[json-rpc] Add option to fail healthcheck if node is lagging
For reliability public fullnode operators will want to run multiple
fullnodes behind a load balancer. However when a new fullnode is started
up it will take a while to sync the chain, and we don't want it serving
requests until it is up to date.

This enhances the healthcheck endpoint to return 503 if the node is
lagging, which will automatically exclude it from the load balancer. Lag
is calculated as the difference in seconds between the local committed
LI and the latest known LI known by state-sync. The threshold is set by
the `max_lag` query param in the healthcheck request.

Test Plan:

    % curl -i 'http://localhost:38457/-/healthy'
    HTTP/1.1 200 OK
    content-type: text/plain; charset=utf-8
    content-length: 13
    date: Fri, 06 Nov 2020 23:13:00 GMT

    libra-node:ok

    % curl -i 'http://localhost:38457/-/healthy?max_lag=60'
    HTTP/1.1 200 OK
    content-type: text/plain; charset=utf-8
    content-length: 16
    date: Fri, 06 Nov 2020 23:13:03 GMT

    libra-node:lag=0

    % curl -i 'http://localhost:38457/-/healthy?max_lag=0'
    HTTP/1.1 503 Service Unavailable
    content-type: text/plain; charset=utf-8
    content-length: 16
    date: Fri, 06 Nov 2020 23:13:06 GMT

    libra-node:lag=0